batch send
At first, this was a difficult change. I tought that it won't matter an additional system call for small sized inputs since we will have to manage a vector that will be on heap. It turns out that even for one input in the vector, it's faster to read/write from heap than doing an aditional sys call.