fix shift, ctrl and alt keys
those keys were not working properly as functional keys since they were pressed and released before the modifier key was pressed.

even though, this introduces a new issue, where if the 'layer' key is released before modifier key, then it keeps being sent as a repeat key. The solution for this will be to track all keys, and release them when the 'layer' key is released, aka CapsLock.